--- src/main/java/site/metacoding/animalprojectfrontend/config/user/UserRepository.java
@@ -1,10 +1,0 @@
-package site.metacoding.animalprojectfrontend.config.user;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.query.Param;
-
-public interface UserRepository extends JpaRepository<User, Integer> {
-    @Query(value = "SELECT * FROM user WHERE username=:username AND password=:password", nativeQuery = true)
-    User aLogin(@Param("username") String username, @Param("password") String password);
-}
