--- src/main/java/site/metacoding/animalprojectfrontend/domain/UserService.java
@@ -1,19 +1,0 @@
-package site.metacoding.animalprojectfrontend.domain;
-
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-import lombok.RequiredArgsConstructor;
-import site.metacoding.animalprojectfrontend.config.user.User;
-import site.metacoding.animalprojectfrontend.config.user.UserRepository;
-
-@RequiredArgsConstructor
-@Service
-public class UserService {
-    private final UserRepository userRepository;
-
-    @Transactional
-    public User name() {
-        
-    }
-}
