+++ src/main/java/site/metacoding/animalprojectfrontend/domain/user/UserRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
 public interface UserRepository extends JpaRepository<User, Integer> {
-    @Query(value = "SELECT * FROM user WHERE username = :username AND password = :password", nativeQuery = true)
-    User mLogin(@Param("username") String username, @Param("password") String password);
+    @Query(value = "SELECT * FROM user WHERE username=:username AND password=:password", nativeQuery = true)
+    User aLogin(@Param("username") String username, @Param("password") String password);
 }
