+++ src/main/java/site/metacoding/animalprojectfrontend/service/UserService.java
@@ -6,7 +6,7 @@ import org.springframework.transaction.annotation.Transactional;
 import lombok.RequiredArgsConstructor;
 import site.metacoding.animalprojectfrontend.domain.user.User;
 import site.metacoding.animalprojectfrontend.domain.user.UserRepository;
-import site.metacoding.animalprojectfrontend.web.api.dto.JoinDto;
+import site.metacoding.animalprojectfrontend.web.api.dto.user.JoinDto;
 import site.metacoding.animalprojectfrontend.web.api.dto.user.LoginDto;
 
 @RequiredArgsConstructor
@@ -20,7 +20,7 @@ public class UserService {
     }
 
     public User 로그인(LoginDto loginDto) {
-        User userEntity = userRepository.mLogin(loginDto.getUsername(), loginDto.getPassword());
+        User userEntity = userRepository.aLogin(loginDto.getUsername(), loginDto.getPassword());
         return userEntity;
     }
 }
