--- src/main/java/site/metacoding/animalprojectfrontend/web/api/dto/JoinDto.java
@@ -1,27 +1,0 @@
-package site.metacoding.animalprojectfrontend.web.api.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import site.metacoding.animalprojectfrontend.domain.user.User;
-
-@AllArgsConstructor
-@NoArgsConstructor
-@Data
-public class JoinDto {
-    private String username;
-    private String password;
-    private String email;
-    private String sex;
-    private String addr;
-
-    public User toEntity() {
-        User user = new User();
-        user.setUsername(this.username);
-        user.setPassword(this.password);
-        user.setEmail(this.email);
-        user.setSex(this.sex);
-        user.setAddr(this.addr);
-        return user;
-    }
-}
