+++ src/main/java/site/metacoding/animalprojectfrontend/web/api/dto/user/JoinDto.java
@@ -3,6 +3,7 @@ package site.metacoding.animalprojectfrontend.web.api.dto.user;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+import site.metacoding.animalprojectfrontend.domain.user.User;
 
 @AllArgsConstructor
 @NoArgsConstructor
@@ -12,4 +13,15 @@ public class JoinDto {
     private String password;
     private String email;
     private String male;
+    private String addr;
+
+    public User toEntity() {
+        User user = new User();
+        user.setUsername(this.username);
+        user.setPassword(this.password);
+        user.setEmail(this.email);
+        user.setMale(this.male);
+        user.setAddr(this.addr);
+        return user;
+    }
 }
