+++ src/main/resources/templates/layout/header.mustache
@@ -31,6 +31,7 @@
 </head>
 
 <body>
+  
   <div class="d-flex justify-content-center">
     <div class="border w-75 bg-success">
       <div class="d-flex justify-content-around mt-4 mb-4">
