+++ src/main/resources/templates/main/joinForm.mustache
@@ -1,83 +1,75 @@
-{{>/layout/header}}
+{{> /layout/header}}
+
 <!-- 컨테이너 시작 -->
-<div class="m_padding_medium mt-3">
+<div class="container mt-3">
 
-  <h1>회원가입</h1>
-  <hr>
-
-  <!-- 회원가입 폼 시작 -->
-  <form class="m_auth_box" novalidate>
+    <!-- 회원가입 폼 시작 -->
+    <form class="m_auth_box" novalidate>
     <div class="form-group">
       <label>아이디</label>
       <input id="username" type="text" class="form-control mt-3" placeholder="username" required>
     </div>
-    <div class="form-group mt-3">
+        <div class="form-group mt-3">
       <label>비밀번호</label>
       <input id="password" type="password" name="password" class="form-control"  title="password" required>
-    </div>
     <div class="form-group mt-3">
       <label>이메일</label>
       <input id="email" type="email" class="form-control" placeholder="name@example.com" required>
-    </div>
     <div class="form-group mt-3">
       <label>성별</label>
-      <select id="sex" class="form-select" required>
+      <select id="male" class="form-select" required>
         <option selected>성별</option>
         <option value="male">남자</option>
         <option value="female">여자</option>
         <option value="female">선택안함</option>
       </select>
     </div>
-    <div class="form-group mt-3">
+        <div class="form-group mt-3">
       <label>주소</label>
       <input id="addr" type="text" class="form-control" minlength="10" data-v-message="Please enter minimum 10 characters"
         required>
     </div>
-    <!-- <div class="form-group mt-5"> -->
-      <button id="btn-join" type="submit" class="btn bg-success text-light" value="회원가입">회원가입</button>
-    <!-- </div> -->
-  </form>
-  <!-- 회원가입 폼 끝 -->
+        <input id="btn-join" type="button" class="btn btn-primary" value="회원가입">
+    </form>
+    <!-- 회원가입 폼 끝 -->
+
 </div>
 
 <script>
-  $("#btn-join").click(() => {
-      join();
-  });
+    $("#btn-join").click(() => {
+    join();
+});
+async function join() {
+    // (1) username, password, email, addr 을 찾아서 오브젝트로 만든다.
+    let joinDto = {
+        username: $("#username").val(),
+        password: $("#password").val(),
+        email: $("#email").val(),
+        male: $("#male option:selected").val(),
+        addr: $("#addr").val()
+    }
 
-  async function join() {
-
-
-
-      let joinDto = {
-          username: $("#username").val(),
-          password: $("#password").val(),
-          email: $("#email").val(),
-          sex:$("#sex").val(),
-          addr: $("#addr").val()
-      }
-
-      let response = fetch("/api/join", {
+    // (2) fetch 요청한다. (json으로 변환해서)
+    let response = await fetch("/join", {
         method: "POST",
         body: JSON.stringify(joinDto),
         headers: {
-          'Content-Type': 'application/json; charset=utf-8'
-        }
-      });
-      
-      let responseParse = response.json();
-      console.log(responseParse);
+            'Content-Type': 'application/json; charset=utf-8'
+        },
+    });
+    let responseParse = await response.json();
+    console.log(responseParse);
 
-      console.log(responseParse.code)
-      if (responseParse.code == 1) {
-          alert("회원가입완료");
-          location.href = "/loginForm";
-      } else {
-          alert('회원가입실패');
-      }
-  }
+        // (3) 회원가입이 잘되면 알림창 띄우고 로그인 페이지로 이동한다.
+    if (responseParse.code == 1) {
+        alert("회원가입완료");
+        location.href = "loginForm";
+    } else {
+        alert('회원가입실패');
+    }
+}
+
 </script>
 
 <!-- 컨테이너 끝 -->
-
-{{>/layout/footer}}
+{{> /layout/footer}}
